Undo image encoding workaround with new ebooklib releases
The new 0.16 release of ebooklib has fixed the bug we were
"workarounding", and the workaround now causes encoding issues.

The plan was to remove the workaround when the fix was released, but
many systems won't get updated ebooklib for a while, so instead of
removing the workaround completelly, I've made it work based on obooklib
version number, so it can work on both older and newer releases, and
postpone the workaround removal for a while.

Fixes #7