Remove an unused CSS file
I had intended to include a custom CSS to style the body of all chapters, but
never got to do it and, over time, I've come to think this was a bad idea.

First, it would impose my preferences on everybody.
Second, it's a waste of time since most readers allow some degree of
customization that overrides those styles.